ENG-5470: Summarize the changes made by pull (#242)
This also switches the default behavior to preserve .fsl files that are
not part of the schema pulled from the remote.

+++ src/commands/schema/pull.js
@@ -2,14 +2,14 @@ const SchemaCommand = require("../../lib/schema-command.js");
 const fetch = require("node-fetch");
 const fs = require("fs");
 const path = require("path");
-const { Flags } = require("@oclif/core");
+const { Flags, ux } = require("@oclif/core");
 
 class PullSchemaCommand extends SchemaCommand {
   static flags = {
     ...SchemaCommand.flags,
-    retain: Flags.boolean({
+    delete: Flags.boolean({
       description:
-        "Retain .fsl files in the target directory that are not part of the database schema",
+        "Delete .fsl files in the target directory that are not part of the database schema",
       default: false,
     }),
     // NB: Required as a flag because it will become optional eventually,
@@ -20,12 +20,27 @@ class PullSchemaCommand extends SchemaCommand {
     }),
   };
 
+  async confirm() {
+    const resp = await ux.prompt("Accept the changes?", {
+      default: "no",
+    });
+    if (["yes", "y"].includes(resp.toLowerCase())) {
+      return true;
+    }
+    if (["no", "n"].includes(resp.toLowerCase())) {
+      return false;
+    }
+    console.log("Please type 'yes' or 'no'");
+    return this.confirm();
+  }
+
   async run() {
     const { urlbase, secret } = await this.fetchsetup();
 
     const dir = this.flags.dir;
 
     try {
+      // Gather remote schema files to download.
       const filesres = await fetch(`${urlbase}/schema/1/files`, {
         method: "GET",
         headers: { AUTHORIZATION: `Bearer ${secret}` },
@@ -40,29 +55,63 @@ class PullSchemaCommand extends SchemaCommand {
         .filter((name) => name.endsWith(".fsl"))
         .sort();
 
-      // All of the below could be parallelized if necessary.
-      if (!this.flags.retain) {
-        // Delete all .fsl files (not directories).
-        for (const deleteme of fs.readdirSync(dir)) {
-          const fp = path.join(dir, deleteme);
-          if (deleteme.endsWith(".fsl") && !fs.statSync(fp).isDirectory()) {
-            fs.unlinkSync(fp);
-          }
+      // Gather local .fsl files to overwrite or delete.
+      const existing = this.gather(dir);
+
+      // Summarize file changes.
+      const adds = [];
+      const overwrites = [];
+      for (const fn of filenames) {
+        if (existing.includes(fn)) {
+          overwrites.push(fn);
+        } else {
+          adds.push(fn);
+        }
+      }
+      const deletes = [];
+      for (const fn of existing) {
+        if (!filenames.includes(fn)) {
+          deletes.push(fn);
+        }
+      }
+      deletes.sort();
+
+      console.log("Pull makes the following changes:");
+      if (this.flags.delete) {
+        for (const deleteme of deletes) {
+          console.log(`delete:    ${deleteme}`);
+        }
+      }
+      for (const add of adds) {
+        console.log(`add:       ${add}`);
+      }
+      for (const overwrite of overwrites) {
+        console.log(`overwrite: ${overwrite}`);
+      }
+
+      if (this.flags.delete) {
+        // Delete extra .fsl files.
+        for (const deleteme of deletes) {
+          fs.unlinkSync(path.join(dir, deleteme));
         }
       }
 
-      for (const filename of filenames) {
-        const fileres = await fetch(`${urlbase}/schema/1/files/${filename}`, {
-          method: "GET",
-          headers: { AUTHORIZATION: `Bearer ${secret}` },
-        });
-        const filejson = await fileres.json();
-        if (filejson.error) {
-          this.error(filejson.error.message);
+      if (await this.confirm()) {
+        for (const filename of filenames) {
+          const fileres = await fetch(`${urlbase}/schema/1/files/${filename}`, {
+            method: "GET",
+            headers: { AUTHORIZATION: `Bearer ${secret}` },
+          });
+          const filejson = await fileres.json();
+          if (filejson.error) {
+            this.error(filejson.error.message);
+          }
+          const fp = path.join(dir, filename);
+          fs.mkdirSync(path.dirname(fp), { recursive: true });
+          fs.writeFileSync(fp, filejson.content);
         }
-        const fp = path.join(dir, filename);
-        fs.mkdirSync(path.dirname(fp), { recursive: true });
-        fs.writeFileSync(fp, filejson.content);
+      } else {
+        this.log("Change cancelled");
       }
     } catch (err) {
       this.error(err);
@@ -73,8 +122,6 @@ class PullSchemaCommand extends SchemaCommand {
 PullSchemaCommand.description =
   "Pull a database schema's .fsl files into a directory";
 
-PullSchemaCommand.examples = [
-  "$ fauna schema pull --dir schemas/myschema --retain",
-];
+PullSchemaCommand.examples = ["$ fauna schema pull --dir schemas/myschema"];
 
 module.exports = PullSchemaCommand;
